backend/pyproject.toml
[project]
name = "bdayateam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.9.0",
    "django-imagekit>=6.0.0",
    "django-phonenumber-field[phonenumberslite]>=8.3.0",
    "django-redis>=6.0.0",
    "django-solo>=2.4.0",
    "django-unfold>=0.72.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular[sidecar]>=0.29.0",
    "gevent>=25.9.1",
    "graphene-django>=3.2.3",
    "graphene-django-optimizer>=0.10.0",
    "graphene-file-upload>=1.3.0",
    "gunicorn>=23.0.0",
    "huey>=2.6.0",
    "msgspec>=0.20.0",
    "pillow>=12.0.0",
    "psycogreen>=1.0.2",
    "psycopg[binary,pool]>=3.3.1",
    "python-dotenv>=1.2.1",
    "redis[hiredis]>=7.1.0",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "django-er-diagram>=3.4.0",
    "faker>=39.0.0",
    "locust>=2.43.0",
    "locust-plugins>=5.0.0",
    "memory-profiler>=0.61.0",
    "ps-mem>=3.14",
    "psutil>=7.2.0",
    "requests>=2.32.5",
]
